refactor: move lookup data models to new package
