+++ shared/src/commonMain/kotlin/no/ordbokene/model/json/lookup/ArticleBody.kt
@@ -1,4 +1,4 @@
-package no.ordbokene.model.json
+package no.ordbokene.model.json.lookup
 
 import kotlinx.serialization.Serializable
