+++ shared/src/commonMain/kotlin/no/ordbokene/model/json/lookup/ContentNode.kt
@@ -1,6 +1,6 @@
 @file:OptIn(ExperimentalSerializationApi::class)
 
-package no.ordbokene.model.json
+package no.ordbokene.model.json.lookup
 
 import kotlinx.serialization.ExperimentalSerializationApi
 import kotlinx.serialization.Polymorphic
