+++ shared/src/commonMain/kotlin/no/ordbokene/model/json/lookup/EtymologyNode.kt
@@ -1,4 +1,4 @@
-package no.ordbokene.model.json
+package no.ordbokene.model.json.lookup
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
